refactor(cmd): extract up command into cmd/up subpackage
Move all up-related source and test files from cmd/ into cmd/up/ as
package up. This follows the existing subpackage pattern (cmd/machine/,
cmd/ide/, cmd/flags/) and isolates the up command's implementation for
better maintainability and navigability.

- cmd/up.go, up_flags.go, up_validate.go, up_client.go → cmd/up/
- cmd/up_test.go, up_update_remote_user_uid_test.go → cmd/up/
- Updated cmd/root.go to import cmd/up and call up.NewUpCmd()
- Updated cmd/minor_flags_test.go and cmd/flag_aliases_test.go to
  import cmd/up for NewUpCmd references
