+++ cmd/flag_aliases_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/devsy-org/devsy/cmd/flags"
+	"github.com/devsy-org/devsy/cmd/up"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -22,7 +23,7 @@ const (
 )
 
 func TestUpCmd_ConfigAlias(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	err := upCmd.ParseFlags([]string{flagConfig, ".devcontainer/custom.json"})
 	require.NoError(t, err)
 
@@ -50,7 +51,7 @@ func TestGlobalFlags_LogFormatAlias(t *testing.T) {
 }
 
 func TestConfigAlias_IsHidden(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	f := upCmd.Flags().Lookup("config")
 	require.NotNil(t, f)
 	assert.True(t, f.Hidden, flagConfig+" alias should be hidden")
@@ -71,7 +72,7 @@ func TestConfigAlias_E2E(t *testing.T) {
 }
 
 func TestUpCmd_OverrideConfigAlias(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	err := upCmd.ParseFlags([]string{flagOverrideConfig, "/tmp/override.json"})
 	require.NoError(t, err)
 
@@ -81,14 +82,14 @@ func TestUpCmd_OverrideConfigAlias(t *testing.T) {
 }
 
 func TestOverrideConfigAlias_IsHidden(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	f := upCmd.Flags().Lookup("override-config")
 	require.NotNil(t, f)
 	assert.True(t, f.Hidden, flagOverrideConfig+" alias should be hidden")
 }
 
 func TestUpCmd_DotfilesRepositoryAlias(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	err := upCmd.ParseFlags([]string{flagDotfilesRepository, "https://github.com/user/dotfiles"})
 	require.NoError(t, err)
 
@@ -98,14 +99,14 @@ func TestUpCmd_DotfilesRepositoryAlias(t *testing.T) {
 }
 
 func TestDotfilesRepositoryAlias_IsHidden(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	f := upCmd.Flags().Lookup("dotfiles-repository")
 	require.NotNil(t, f)
 	assert.True(t, f.Hidden, flagDotfilesRepository+" alias should be hidden")
 }
 
 func TestUpCmd_RemoveExistingContainerAlias(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	err := upCmd.ParseFlags([]string{flagRemoveExistingContainer})
 	require.NoError(t, err)
 
@@ -115,7 +116,7 @@ func TestUpCmd_RemoveExistingContainerAlias(t *testing.T) {
 }
 
 func TestRemoveExistingContainerAlias_IsHidden(t *testing.T) {
-	upCmd := NewUpCmd(&flags.GlobalFlags{})
+	upCmd := up.NewUpCmd(&flags.GlobalFlags{})
 	f := upCmd.Flags().Lookup("remove-existing-container")
 	require.NotNil(t, f)
 	assert.True(t, f.Hidden, flagRemoveExistingContainer+" alias should be hidden")
