+++ cmd/up/up.go
@@ -1,4 +1,4 @@
-package cmd
+package up
 
 import (
 	"context"
