+++ cmd/up/up_update_remote_user_uid_test.go
@@ -1,4 +1,4 @@
-package cmd
+package up
 
 import (
 	"testing"
