+++ cmd/up/up_validate.go
@@ -1,4 +1,4 @@
-package cmd
+package up
 
 import (
 	"fmt"
